Adding numba as optional dependency
