+++ pyproject.toml
@@ -24,10 +24,12 @@ classifiers = [
 dependencies = [
     "astropy>=5.2.2",
     "matplotlib>=3.7.5",
-    "mpmath>=1.3.0",
     "numpy>=1.24.4",
     "scipy>=1.10.1",
 ]
+
+[project.optional-dependencies]
+numba = ["numba>=0.56.0"]
 
 [project.scripts]
 NE2025p = "mwprop.cli.NE2025p:main"
