--- setup.cfg
@@ -1,33 +1,0 @@
-[metadata]
-name = mwprop
-version = 2.0.0
-author = James Cordes and Stella Ocker
-author_email = stella.ocker@gmail.com
-description = A python package for NE2025 and NE2001
-long_description = file: README.md
-long_description_content_type = text/markdown
-url = https://github.com/stella-ocker/mwprop
-project_urls =
-    Bug Tracker = https://github.com/stella-ocker/mwprop/issues
-classifiers =
-    Programming Language :: Python :: 3
-    License :: OSI Approved :: GNU General Public License (GPL)
-    Operating System :: OS Independent
-
-[options]
-package_dir =
-    = src
-packages = mwprop,mwprop.nemod,mwprop.scattering_functions
-include_package_data = True
-python_requires = >=3.6
-scripts =
-    bin/NE2025p.py
-    bin/NE2001p.py
-    bin/los_diagnostics.py
-    bin/test_ne2025p.py
-
-[options.package_data]
-* = *.txt,params/*
-
-[options.packages.find]
-where = src
